+++ code/parachain/runtime/primitives/src/currency.rs
@@ -371,70 +371,3 @@ mod ops {
 		}
 	}
 }
-
-#[allow(clippy::large_enum_variant)]
-#[derive(RuntimeDebug, Decode, Encode, Clone, PartialEq, Eq, TypeInfo, MaxEncodedLen)]
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
-pub enum ForeignAssetId {
-	Xcm(VersionedMultiLocation),
-	IbcIcs20(PrefixedDenom),
-}
-
-#[derive(
-	Ord, PartialOrd, RuntimeDebug, Decode, Encode, Clone, PartialEq, Eq, TypeInfo, MaxEncodedLen,
-)]
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
-pub enum VersionedMultiLocation {
-	#[codec(index = 3)]
-	V3(v3::MultiLocation),
-}
-
-impl From<VersionedMultiLocation> for ForeignAssetId {
-	fn from(this: VersionedMultiLocation) -> Self {
-		Self::Xcm(this)
-	}
-}
-
-impl From<PrefixedDenom> for ForeignAssetId {
-	fn from(this: PrefixedDenom) -> Self {
-		Self::IbcIcs20(this)
-	}
-}
-
-type InnerDenom = ibc_rs_scale::applications::transfer::PrefixedDenom;
-
-#[derive(Debug, Decode, Clone, PartialEq, Eq, TypeInfo)]
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
-#[cfg_attr(feature = "std", serde(transparent))]
-pub struct PrefixedDenom(pub InnerDenom);
-
-impl FromStr for PrefixedDenom {
-	type Err = DispatchError;
-	fn from_str(s: &str) -> Result<Self, DispatchError> {
-		InnerDenom::from_str(s)
-			.map_err(|_| DispatchError::Other("PrefixedDenom parse failed"))
-			.map(Self)
-	}
-}
-
-impl WrapperTypeEncode for PrefixedDenom {}
-impl EncodeLike for PrefixedDenom {}
-impl core::ops::Deref for PrefixedDenom {
-	type Target = InnerDenom;
-
-	fn deref(&self) -> &Self::Target {
-		&self.0
-	}
-}
-
-impl From<InnerDenom> for PrefixedDenom {
-	fn from(this: InnerDenom) -> Self {
-		Self(this)
-	}
-}
-
-impl MaxEncodedLen for PrefixedDenom {
-	fn max_encoded_len() -> usize {
-		2048
-	}
-}
